move FusedCopy type and logic into its own submodule
